Refactor TurretPicker component and adjust imports
This commit refactors and relocates the TurretPicker component from 'features' to 'components' directory. This has been done to reduce the coupling between the TurretPicker and the Redux store by moving Redux logic up the tree. This makes TurretPicker more reusable and reduces its responsibilities.

The imports and its uses in 'turret-planner', 'getting-started' and other components have been adjusted accordingly.

Additionally, turret clearing functionality has been moved to the 'turret-planner' page, as it aligns more closely with this page's responsibilities.
