--- src/pages/getting-started/desktop/styles.module.css
@@ -1,4 +1,0 @@
-
-.text {
-    text-align: justify;
-}
