--- src/pages/getting-started/mobile/styles.module.css
@@ -1,31 +1,0 @@
-@keyframes ArrowJump {
-    0% {
-        transform: translateY(0rem);
-    }
-    60% {
-        transform: translateY(0rem);
-    }
-    70% {
-        transform: translateY(1rem);
-    }
-    80% {
-        transform: translateY(0rem);
-    }
-    90% {
-        transform: translateY(1rem);
-    }
-    100% {
-        transform: translateY(0rem);
-    }
-}
-
-.arrow {
-    animation-duration: 3s;
-    animation-iteration-count: infinite;
-    animation-timing-function: ease-in-out;
-    animation-name: ArrowJump;
-}
-
-.text {
-    text-align: justify;
-}
